modified:   README.md 	modified:   controllers/productos.controller.js 	modified:   schemas/productos.schema.js

+++ controllers/productos.controller.js
@@ -15,26 +15,31 @@ export const getProductoById = async (req, res) => {
     const { id } = req.params
     const producto = await ProductosService.getProductoById(parseInt(id))
     if (producto == null) {
-        sendResponse({res, message: "Producto no encontrado", data: null, statusCode: 404 })
+        return sendResponse({res, message: "Producto no encontrado", data: null, statusCode: 404 })
     }
-    sendResponse({res, message: "Producto obtenido", data: producto, statusCode: 201 })
+    sendResponse({res, message: "Producto obtenido", data: producto})
 }
 
 export const createProducto = async (req, res) => {
     let producto={...req.body,"fecha_ingreso":new Date().toISOString()}
     const { success, error, data: safeData } = validateproductos(producto)
 
+    const existProducto = await ProductosService.getProductoById(parseInt(safeData.id))
+    if (!existProducto===false) {
+        return sendResponse({ res, message: "EL id ya existe"})
+    }
+
     if (!success) {
         return sendResponse({ res, message: "ocurrio un error", data: error.issues, statusCode: 400 })
     }
     const data = await ProductosService.createProducto(safeData)
 
-    sendResponse({ res, message: "Producto creado con exito", data: data, statusCode: 201 })
+    sendResponse({ res, message: "Producto creado con exito", data: data})
 }
 
 export const updateProducto = async (req, res) => {
     const { id } = req.params
-    let productoActualizado={...req.body,"fecha_ingreso":new Date().toISOString()}
+    let productoActualizado={"id":parseInt(id),...req.body,"fecha_ingreso":new Date().toISOString()}
     const { success, error, data: safeData } = validateproductos(productoActualizado)
 
     if (!success) {
@@ -42,31 +47,31 @@ export const updateProducto = async (req, res) => {
     }
     const productoAnterior=await ProductosService.getProductoById(parseInt(id))
     if (productoAnterior==null){
-        return sendResponse({ res, message: "El producto no existe", statusCode: 400 })   
+        return sendResponse({ res, message: "El producto no existe", statusCode: 404 })   
     }
     const data = await ProductosService.updateProducto(safeData,parseInt(id))
 
-    sendResponse({ res, message: "Producto actualizado con exito", data: data, statusCode: 201 })
+    sendResponse({ res, message: "Producto actualizado con exito", data: data})
 }
 
 export const deleteProducto = async (req, res) => {
     const { id } = req.params
 
-    const producto = await ProductosService.getProductoById(parseInt(id))
-    if (producto === -1) {
-        sendResponse({ res, message: "No existe el producto", statusCode: 404 })
+    const deleteProducto = await ProductosService.getProductoById(parseInt(id))
+    if (!deleteProducto) {
+        return sendResponse({ res, message: "No existe el producto", statusCode: 404 })
     }
 
     await ProductosService.deleteProducto(parseInt(id))
 
-    sendResponse({ res, message: "Producto eliminado con exito", data: producto, statusCode: 201 })
+    sendResponse({ res, message: "Producto eliminado con exito", data: deleteProducto})
 }
 
-export const getProductosDisponibles = async (req, res/*, next*/) => {
+export const getProductosDisponibles = async (req, res) => {
     try {
         const productosDisponibles = await ProductosService.getProductosDisponibles();
         sendResponse({ res, message: "Listado de productos disponibles", data: productosDisponibles})
     } catch (e) {
-        sendResponse({ res, message: e.message, data: null, statusCode: 500 })
+        sendResponse({ res, message: e.message, statusCode: 500 })
     }
 }
